working on createJob, need to test with a request body

+++ src/main/java/com/qu1cksave/qu1cksave_backend/coverletter/RequestCoverLetterDto.java
@@ -3,9 +3,8 @@ package com.qu1cksave.qu1cksave_backend.coverletter;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class RequestCoverLetterDto {
-    private final String fileName;
-
-    private final String mimeType;
+    private final String fileName; // NOT NULLABLE
+    private final String mimeType; // NOT NULLABLE
 
     // From qu1cksave backend Express-version:
     // - bytearray_as_array: number[];  // Changed from Uint8Array to number[]
@@ -13,8 +12,8 @@ public class RequestCoverLetterDto {
     // - The JavaScript Number type is a double-precision 64-bit binary format IEEE 754 value, like double in Java or C#.
     // TODO: (3/29/25) Maybe I can use long?
     //  Though, there's also byte[]
-    //  I'm using long temporarily for now
-    private final long[] byteArrayAsArray;
+    //  I'm using double temporarily for now
+    private final double[] byteArrayAsArray;
 
     // Constructors
     // Need JsonProperty in constructor params so Jackson knows how to
@@ -22,7 +21,7 @@ public class RequestCoverLetterDto {
     public RequestCoverLetterDto(
         @JsonProperty("fileName") String fileName,
         @JsonProperty("mimeType") String mimeType,
-        @JsonProperty("byteArrayAsArray") long[] byteArrayAsArray
+        @JsonProperty("byteArrayAsArray") double[] byteArrayAsArray
     ) {
         this.fileName = fileName;
         this.mimeType = mimeType;
@@ -32,5 +31,5 @@ public class RequestCoverLetterDto {
     // Getters
     public String getFileName() { return fileName; }
     public String getMimeType() { return mimeType; }
-    public long[] getByteArrayAsArray() { return byteArrayAsArray; }
+    public double[] getByteArrayAsArray() { return byteArrayAsArray; }
 }
