+++ src/main/java/com/qu1cksave/qu1cksave_backend/resume/RequestResumeDto.java
@@ -4,9 +4,8 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 
 public class RequestResumeDto {
-    private final String fileName;
-
-    private final String mimeType;
+    private final String fileName; // NOT NULLABLE
+    private final String mimeType; // NOT NULLABLE
 
     // From qu1cksave backend Express-version:
     // - bytearray_as_array: number[];  // Changed from Uint8Array to number[]
@@ -14,16 +13,18 @@ public class RequestResumeDto {
     // - The JavaScript Number type is a double-precision 64-bit binary format IEEE 754 value, like double in Java or C#.
     // TODO: (3/29/25) Maybe I can use long?
     //  Though, there's also byte[]
-    //  I'm using long temporarily for now
-    private final long[] byteArrayAsArray;
+    //  I'm using double temporarily for now
+    //  long can also work
+    private final double[] byteArrayAsArray;
 
     // Constructors
     // Need JsonProperty in constructor params so Jackson knows how to
     //   deserialize
     public RequestResumeDto(
+        // TODO: (5/2/25) Might want to change this to snake_case
         @JsonProperty("fileName") String fileName,
         @JsonProperty("mimeType") String mimeType,
-        @JsonProperty("byteArrayAsArray") long[] byteArrayAsArray
+        @JsonProperty("byteArrayAsArray") double[] byteArrayAsArray
     ) {
         this.fileName = fileName;
         this.mimeType = mimeType;
@@ -33,5 +34,5 @@ public class RequestResumeDto {
     // Getters
     public String getFileName() { return fileName; }
     public String getMimeType() { return mimeType; }
-    public long[] getByteArrayAsArray() { return byteArrayAsArray; }
+    public double[] getByteArrayAsArray() { return byteArrayAsArray; }
 }
